pyproject.toml
[project]
name = "projet-mise-en-prod-3a"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "inflect>=7.5.0",
    "loguru>=0.7.3",
    "pandas>=2.2.3",
    "pyspellchecker>=0.8.2",
    "pyyaml>=6.0.2",
    "s3fs>=2025.3.0",
    "streamlit>=1.40.2",
    "streamlit-extras>=0.5.5",
    "flask",
    "flask_login"
]
